Fix day of week mapping for Sunday
`Date.prototype.getDay()` [1] returns 0 for Sunday. Thus previously on
Sundays the `dayNames` array lookup returned `undefined` and no CSS
class was added to the element.

By moving Sunday ('sun') to the beginning of `dayNames` and using the
`getDay()` index directly we can fix the problem so that a 'sun' CSS
class is correctly added to the element.

Ideally I would've added some tests, but given that I'm not sure we're
going to want to continue to use this plugin and I would need to setup
the whole testing infrastructure for the project, I've decided it's not
worth it at the moment!

[1]: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/getDay